optimized the code and implemented the summerization process under a new function in a new class for more usability
